+++ src/Raven.Server/Documents/Queries/IndexQueryServerSide.cs
@@ -73,10 +73,10 @@ namespace Raven.Server.Documents.Queries
                         result.ExplainScores = (bool)propertyDetails.Value;
                         break;
                     case nameof(PageSize):
-                        result.PageSize = (int)propertyDetails.Value;
+                        result.PageSize = (int)(long)propertyDetails.Value;
                         break;
                     case nameof(Start):
-                        result.Start = (int)propertyDetails.Value;
+                        result.Start = (int)(long)propertyDetails.Value;
                         break;
                     case nameof(ShowTimings):
                         result.ShowTimings = (bool)propertyDetails.Value;
